backend/requirements.txt
Django==5.1.5
djangorestframework==3.15.2
django-cors-headers==4.6.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
bcrypt==4.2.1
requests==2.32.3
python-dotenv==1.0.1
Flask==3.1.0
Flask-Cors==5.0.0
pandas==2.2.3
earthengine-api>=0.1.390
